requirements.txt
torch
torchvision
opencv-python
numpy
pillow
scikit-learn
mtcnn
tqdm
streamlit
mtcnn
matplotlib
